scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.11",
    "fastapi>=0.131.0",
    "langchain-groq>=1.1.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
